# openapi/scheduling-internal-v1.yaml
openapi: 3.0.3
info:
  title: AutoProduction Scheduling Internal API
  version: 1.0.0
  description: Internal API contract for scheduling console, approval workflow, publish/rollback and ops query.
servers:
  - url: /internal/v1
security:
  - bearerAuth: []
...
paths:
  /internal/order-pool:
    get:
      tags: [order]
      summary: Query order pool
      parameters:
        - $ref: '#/components/parameters/Page'
        - $ref: '#/components/parameters/PageSize'
        - in: query
          name: status
          schema: { type: string }
        - in: query
          name: frozen_flag
          schema: { type: integer, enum: [0, 1] }
        - in: query
          name: urgent_flag
          schema: { type: integer, enum: [0, 1] }
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema: { $ref: '#/components/schemas/OrderPoolListResponse' }
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '500': { $ref: '#/components/responses/ServerError' }
  /internal/schedule-versions:
    get:
      tags: [version]
      summary: Query schedule versions
      parameters:
        - $ref: '#/components/parameters/Page'
        - $ref: '#/components/parameters/PageSize'
        - in: query
          name: status
          schema: { type: string }
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ScheduleVersionListResponse' }
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '500': { $ref: '#/components/responses/ServerError' }
  /internal/schedule-versions/{version_no}/tasks:
    get:
      tags: [version]
      summary: Query schedule tasks by version
      parameters:
        - in: path
          name: version_no
          required: true
          schema: { type: string }
        - $ref: '#/components/parameters/Page'
        - $ref: '#/components/parameters/PageSize'
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ScheduleTaskListResponse' }
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
        '500': { $ref: '#/components/responses/ServerError' }
  /internal/schedule-versions/{version_no}/diff:
    get:
      tags: [version]
      summary: Query version diff summary
      parameters:
        - in: path
          name: version_no
          required: true
          schema: { type: string }
        - in: query
          name: compare_with
          required: true
          schema: { type: string }
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema: { $ref: '#/components/schemas/VersionDiffResponse' }
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
        '500': { $ref: '#/components/responses/ServerError' }
  /internal/dispatch-commands:
    get:
      tags: [dispatch]
      summary: Query dispatch commands
      parameters:
        - $ref: '#/components/parameters/Page'
        - $ref: '#/components/parameters/PageSize'
        - in: query
          name: command_type
          schema: { type: string }
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema: { $ref: '#/components/schemas/DispatchCommandListResponse' }
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '500': { $ref: '#/components/responses/ServerError' }
    post:
      tags: [dispatch]
      summary: Create dispatch command
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/DispatchCommandRequest' }
      responses:
        '202':
          description: Accepted
          content:
            application/json:
              schema: { $ref: '#/components/schemas/AcceptedResponse' }
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '409': { $ref: '#/components/responses/Conflict' }
        '500': { $ref: '#/components/responses/ServerError' }
  /internal/dispatch-commands/{command_id}/approvals:
    post:
      tags: [dispatch]
      summary: Approve or reject dispatch command
      parameters:
        - in: path
          name: command_id
          required: true
          schema: { type: string }
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/DispatchApprovalRequest' }
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ActionResponse' }
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
        '409': { $ref: '#/components/responses/Conflict' }
        '500': { $ref: '#/components/responses/ServerError' }
  /internal/schedule-versions/{version_no}/publish:
    post:
      tags: [version]
      summary: Publish schedule version
      parameters:
        - in: path
          name: version_no
          required: true
          schema: { type: string }
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/VersionActionRequest' }
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ActionResponse' }
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
        '409': { $ref: '#/components/responses/Conflict' }
        '500': { $ref: '#/components/responses/ServerError' }
  /internal/schedule-versions/{version_no}/rollback:
    post:
      tags: [version]
      summary: Rollback schedule version
      parameters:
        - in: path
          name: version_no
          required: true
          schema: { type: string }
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/VersionActionRequest' }
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ActionResponse' }
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
        '409': { $ref: '#/components/responses/Conflict' }
        '500': { $ref: '#/components/responses/ServerError' }
  /internal/alerts:
    get:
      tags: [alert]
      summary: Query alert events
      parameters:
        - $ref: '#/components/parameters/Page'
        - $ref: '#/components/parameters/PageSize'
        - in: query
          name: status
          schema: { type: string }
        - in: query
          name: severity
          schema: { type: string }
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema: { $ref: '#/components/schemas/AlertEventListResponse' }
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '500': { $ref: '#/components/responses/ServerError' }
  /internal/alerts/{alert_id}/ack:
    post:
      tags: [alert]
      summary: Acknowledge alert event
      parameters:
        - in: path
          name: alert_id
          required: true
          schema: { type: string }
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/AlertActionRequest' }
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ActionResponse' }
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
        '409': { $ref: '#/components/responses/Conflict' }
        '500': { $ref: '#/components/responses/ServerError' }
  /internal/alerts/{alert_id}/close:
    post:
      tags: [alert]
      summary: Close alert event
      parameters:
        - in: path
          name: alert_id
          required: true
          schema: { type: string }
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/AlertActionRequest' }
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ActionResponse' }
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
        '409': { $ref: '#/components/responses/Conflict' }
        '500': { $ref: '#/components/responses/ServerError' }
  /internal/replan-jobs:
    post:
      tags: [replan]
      summary: Trigger replan job
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/ReplanJobRequest' }
      responses:
        '202':
          description: Accepted
          content:
            application/json:
              schema: { $ref: '#/components/schemas/AcceptedResponse' }
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '409': { $ref: '#/components/responses/Conflict' }
        '500': { $ref: '#/components/responses/ServerError' }
  /internal/replan-jobs/{job_no}:
    get:
      tags: [replan]
      summary: Query replan job
      parameters:
        - in: path
          name: job_no
          required: true
          schema: { type: string }
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ReplanJobResponse' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
        '500': { $ref: '#/components/responses/ServerError' }
  /internal/audit-logs:
    get:
      tags: [audit]
      summary: Query audit logs
      parameters:
        - $ref: '#/components/parameters/Page'
        - $ref: '#/components/parameters/PageSize'
        - in: query
          name: entity_type
          schema: { type: string }
        - in: query
          name: request_id
          schema: { type: string }
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema: { $ref: '#/components/schemas/AuditLogListResponse' }
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '500': { $ref: '#/components/responses/ServerError' }
  /internal/integration/inbox:
    get:
      tags: [integration]
      summary: Query integration inbox messages
      parameters:
        - $ref: '#/components/parameters/Page'
        - $ref: '#/components/parameters/PageSize'
        - in: query
          name: status
          schema: { type: string }
        - in: query
          name: topic
          schema: { type: string }
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema: { $ref: '#/components/schemas/IntegrationMessageListResponse' }
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '500': { $ref: '#/components/responses/ServerError' }
  /internal/integration/outbox:
    get:
      tags: [integration]
      summary: Query integration outbox messages
      parameters:
        - $ref: '#/components/parameters/Page'
        - $ref: '#/components/parameters/PageSize'
        - in: query
          name: status
          schema: { type: string }
        - in: query
          name: topic
          schema: { type: string }
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema: { $ref: '#/components/schemas/IntegrationMessageListResponse' }
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '500': { $ref: '#/components/responses/ServerError' }
  /internal/integration/outbox/{message_id}/retry:
    post:
      tags: [integration]
      summary: Retry integration outbox message
      parameters:
        - in: path
          name: message_id
          required: true
          schema: { type: string }
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/OutboxRetryRequest' }
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ActionResponse' }
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
        '500': { $ref: '#/components/responses/ServerError' }
  /internal/simulation/state:
    get:
      tags: [simulation]
      summary: Query simulation state
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema: { $ref: '#/components/schemas/SimulationStateResponse' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '500': { $ref: '#/components/responses/ServerError' }
  /internal/simulation/events:
    get:
      tags: [simulation]
      summary: Query simulation events
      parameters:
        - $ref: '#/components/parameters/Page'
        - $ref: '#/components/parameters/PageSize'
        - in: query
          name: event_type
          schema: { type: string }
        - in: query
          name: event_date
          schema: { type: string, format: date }
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema: { $ref: '#/components/schemas/SimulationEventListResponse' }
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '500': { $ref: '#/components/responses/ServerError' }
  /internal/simulation/reset:
    post:
      tags: [simulation]
      summary: Reset simulation state
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/SimulationResetRequest' }
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema: { $ref: '#/components/schemas/SimulationStateResponse' }
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '500': { $ref: '#/components/responses/ServerError' }
  /internal/simulation/run:
    post:
      tags: [simulation]
      summary: Run day-forward simulation
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/SimulationRunRequest' }
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema: { $ref: '#/components/schemas/SimulationRunResponse' }
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '500': { $ref: '#/components/responses/ServerError' }
  /internal/simulation/manual/add-production-order:
    post:
      tags: [simulation]
      summary: Add one manual production order to main system
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/ManualSimulationActionRequest' }
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ManualAddProductionOrderResponse' }
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '500': { $ref: '#/components/responses/ServerError' }
  /internal/simulation/manual/advance-day:
    post:
      tags: [simulation]
      summary: Advance one day in main system manual simulation
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/ManualSimulationActionRequest' }
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema: { $ref: '#/components/schemas/SimulationRunResponse' }
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '500': { $ref: '#/components/responses/ServerError' }
  /internal/simulation/manual/reset:
    post:
      tags: [simulation]
      summary: Reset main system to pre-manual-simulation snapshot
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/ManualSimulationActionRequest' }
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema: { $ref: '#/components/schemas/SimulationStateResponse' }
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '500': { $ref: '#/components/responses/ServerError' }
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
    Page:
      in: query
      name: page
      required: false
      schema:
        type: integer
        minimum: 1
        default: 1
    PageSize:
      in: query
      name: page_size
      required: false
      schema:
        type: integer
        minimum: 1
        maximum: 1000
        default: 200
  responses:
    BadRequest:
      description: Invalid request
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ErrorResponse' }
    Unauthorized:
      description: Authentication failed
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ErrorResponse' }
    Forbidden:
      description: Permission denied
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ErrorResponse' }
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ErrorResponse' }
    Conflict:
      description: Business conflict
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ErrorResponse' }
    ServerError:
      description: Internal server error
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ErrorResponse' }
  schemas:
...
    AcceptedResponse:
      type: object
      required: [request_id, accepted]
      properties:
        request_id: { type: string }
        accepted: { type: boolean }
        message: { type: string }
...
    VersionDiffResponse:
      type: object
      required: [request_id, base_version_no, compare_version_no]
      properties:
        request_id: { type: string }
        base_version_no: { type: string }
        compare_version_no: { type: string }
        changed_order_count: { type: integer }
        changed_task_count: { type: integer }
        delayed_order_delta: { type: integer }
        overloaded_process_delta: { type: integer }
...
    DispatchCommandRequest:
      type: object
      required: [request_id, command_type, target_order_no, target_order_type, effective_time, reason, created_by]
      properties:
        request_id: { type: string }
        command_type: { type: string, enum: [INSERT, LOCK, UNLOCK, PRIORITY, FREEZE, UNFREEZE] }
        target_order_no: { type: string }
        target_order_type: { type: string, enum: [sales, production] }
        effective_time: { type: string, format: date-time }
        reason: { type: string }
        created_by: { type: string }
    DispatchApprovalRequest:
      type: object
      required: [request_id, approver, decision, decision_time]
      properties:
        request_id: { type: string }
        approver: { type: string }
        decision: { type: string, enum: [APPROVED, REJECTED] }
        decision_reason: { type: string }
        decision_time: { type: string, format: date-time }
...
    OrderPoolListResponse:
      allOf:
        - $ref: '#/components/schemas/BaseListResponse'
        - type: object
          properties:
            items:
              type: array
              items: { $ref: '#/components/schemas/OrderPoolItem' }
    ScheduleVersionListResponse:
      allOf:
        - $ref: '#/components/schemas/BaseListResponse'
        - type: object
          properties:
            items:
              type: array
              items: { $ref: '#/components/schemas/ScheduleVersionItem' }
    ScheduleTaskListResponse:
      allOf:
        - $ref: '#/components/schemas/BaseListResponse'
        - type: object
          properties:
            items:
              type: array
              items: { $ref: '#/components/schemas/ScheduleTaskItem' }
    DispatchCommandListResponse:
      allOf:
        - $ref: '#/components/schemas/BaseListResponse'
        - type: object
          properties:
            items:
              type: array
              items: { $ref: '#/components/schemas/DispatchCommandItem' }
...